West Germanic update 4
im done modding for tonight. Frankreich is mostly finished, we have a
few more things to do namely ideas dynasty and king names, tech,
religion, and some other housekeeping. Frisia is in the game too and is
in pretty much the same state.

Accidentally borked the cultures someone can try to fix it, or i will
problaby try to fix it later.