Flag Bug Fix
That previous commit actually did fix the flag bug, I just was an idiot
and didn't test it properly. =P

Turns out the game doesn't like it if you have flags that aren't being
used in the main flag folder, so if you find a cool flag put it in
another folder please. I think we might be having a similar problem in
the religions file, I will look at that later.