Gaul Culture and religion
ok, fixed up gaul so it is no longer just one culture and one religion,
Insular is still prevalent in the north, and brittany, the only nation
that is now arab-breton is the one holy order thing we were thinking
about having pretty much everywhere else is breton or Aquitanian, the
dominate religion is islam, but insular still exists, but with most of
the insular practicing peoples inside either a Qusayyad lands or in
Frankish Territory.

so yeah, this pretty much leaves only Civ'ed to fix the nations.

Expect European cultural unions, formable nations, and revolters,
soon...ish(tm)